added the recursive array schema generator
